+++ src/main/java/com/kospot/kospot/domain/coordinate/service/CoordinateExcelService.java
@@ -1,0 +1,5 @@
+package com.kospot.kospot.domain.coordinate.service;
+
+public interface CoordinateExcelService {
+
+}
